+++ examples/queries/retailer_2/retailer_2-Q1.sql
@@ -10,7 +10,7 @@ FROM FILE './datasets/retailer_unordered/Location.tbl' LINE DELIMITED CSV(delimi
 CREATE STREAM  Census(zip int, population int, white int, asian int, pacific int, blackafrican int, medianage double, occupiedhouseunits int, houseunits int, families int, households int, husbwife int, males int, females int, householdschildren int, hispanic int)
 FROM FILE './datasets/retailer_unordered/Census.tbl' LINE DELIMITED CSV(delimiter := '|');
 
-CREATE STREAM  Weather(locn int, dateid int, rain byte, snow byte, maxtemp int, mintemp int, meanwind double, thunder byte)
+CREATE STREAM  Weather(locn int, dateid int, rain int, snow int, maxtemp int, mintemp int, meanwind double, thunder int)
 FROM FILE './datasets/retailer_unordered/Weather.tbl' LINE DELIMITED CSV(delimiter := '|');
 
 
@@ -19,6 +19,6 @@ SELECT SUM(
     [lift<1>: RingFactorizedRelation<[1,int]>](zip) *
     [lift<2>: RingFactorizedRelation<[2,int,int,int,int,int,double,double,double,double,double,double,double,double]>](rgn_cd, clim_zn_nbr, tot_area_sq_ft, sell_area_sq_ft, avghhi, supertargetdistance, supertargetdrivetime, targetdistance, targetdrivetime, walmartdistance, walmartdrivetime, walmartsupercenterdistance, walmartsupercenterdrivetime) *
     [lift<15>: RingFactorizedRelation<[15,int,int,int,int,int,double,int,int,int,int,int,int,int,int,int]>](population, white, asian, pacific, blackafrican, medianage, occupiedhouseunits, houseunits, families, households, husbwife, males, females, householdschildren, hispanic) *
-    [lift<30>: RingFactorizedRelation<[30,int,byte,byte,int,int,double,byte]>](dateid,rain, snow, maxtemp, mintemp, meanwind, thunder)
+    [lift<30>: RingFactorizedRelation<[30,int,int,int,int,int,double,int]>](dateid,rain, snow, maxtemp, mintemp, meanwind, thunder)
 )
 FROM Location NATURAL JOIN  Census NATURAL JOIN  Weather;
